// src/app/shared/models/enums.ts
export enum Role {
  ADMIN = 'ADMIN',
  CHECKIN = 'CHECKIN',
}

export enum PlanType {
  PAID = 'PAID',
  SCHOLARSHIP = 'SCHOLARSHIP',
}

export enum PaymentMethod {
  CARD = 'CARD',
  SCHOLARSHIP = 'SCHOLARSHIP',
}

export enum FinancialStatus {
  PENDING = 'PENDING',
  ACTIVE = 'ACTIVE',
  OVERDUE = 'OVERDUE',
  CANCELLED = 'CANCELLED',
  EXEMPT = 'EXEMPT',
}